requirements.txt
python-telegram-bot
pytest
Pillow